Omega refactor
