--- src/components/elements/Animator/Animator.module.scss
@@ -1,130 +1,0 @@
-.animator {
-	position: relative;
-	width: 100%;
-	height: 100%;
-	max-width: 100%;
-	max-height: 100%;
-	display: flex;
-	flex-direction: column;
-}
-.animatorOuterImageContainer {
-	overflow: hidden;
-	flex-grow: 1;
-}
-.animatorInnerImageContainer {
-	position: relative;
-	width: 100%;
-	height: 100%;
-	max-width: 100%;
-	max-height: 100%;
-}
-.dataTooltip {
-	position: absolute;
-	pointer-events: none;
-	z-index: 100;
-
-	&.bottom-right {
-		transform: translate(10px, 10px);
-	}
-
-	&.bottom-left {
-		transform: translate(calc(-100% - 10px), 10px);
-	}
-
-	&.top-right {
-		transform: translate(10px, calc(-100% - 10px));
-	}
-
-	&.top-left {
-		transform: translate(calc(-100% - 10px), calc(-100% - 10px));
-	}
-
-	.tooltipContent {
-		background-color: rgba(0, 0, 0, 0.7);
-		color: white;
-		padding: 8px 12px;
-		border-radius: 4px;
-		font-size: 14px;
-		min-width: 120px;
-
-		p {
-			margin: 4px 0;
-		}
-		.loadingIndicator {
-			font-style: italic;
-			opacity: 0.7;
-			margin: 4px 0;
-		}
-
-		.readoutData {
-			margin-top: 8px;
-			border-top: 1px solid rgba(255, 255, 255, 0.3);
-			padding-top: 6px;
-		}
-	}
-}
-.controlsContainer {
-	flex-shrink: 0;
-	flex-grow: 0;
-	padding: 0 2px;
-	width: 100%;
-	min-width: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	background: var(--color-white-grey18);
-}
-.controls {
-	width: 70%;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	gap: 20px;
-	@media screen and (max-width: 700px) {
-		width: 95%;
-		gap: 10px;
-		flex-wrap: wrap;
-		justify-content: space-between;
-		& > div:first-child {
-			width: 100%;
-			min-width: 100%;
-			margin-bottom: -10px;
-		}
-	}
-}
-.zoomControls {
-	position: absolute;
-	top: 2px;
-	left: 2px;
-	display: flex;
-	flex-direction: column;
-	gap: 2px;
-	button {
-		position: relative;
-		border-radius: 4px;
-		display: block;
-		background-color: var(--color-grey17);
-		outline: none;
-		border: none;
-		padding: 4px;
-		width: 30px;
-		font-size: 15px;
-		text-align: center;
-		height: 30px;
-		color: var(--color-grey5);
-		transition: var(--themeChange);
-		cursor: pointer;
-		&:hover {
-			color: var(--color-white);
-		}
-	}
-}
-.animatorImagesContainer {
-	position: relative;
-	& > div {
-		position: absolute;
-	}
-	& > div:nth-child(1) {
-		position: relative;
-	}
-}
