+++ src/components/elements/Animator/AnimatorControls/AnimatorControls.module.scss
@@ -1,0 +1,29 @@
+.controlsContainer {
+  flex-shrink: 0;
+  flex-grow: 0;
+  padding: 0 2px;
+  width: 100%;
+  min-width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: var(--color-white-grey18);
+}
+.controls {
+  width: 70%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+  @media screen and (max-width: 700px) {
+    width: 95%;
+    gap: 10px;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    & > div:first-child {
+      width: 100%;
+      min-width: 100%;
+      margin-bottom: -10px;
+    }
+  }
+}
