+++ src/components/elements/Animator/AnimatorImageSizer/AnimatorImageSizer.module.scss
@@ -1,0 +1,18 @@
+.animatorImageSizer {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  max-width: 100%;
+  max-height: 100%;
+}
+.animatorImagesContainer {
+  position: relative;
+  & > div {
+    position: absolute;
+    pointer-events: none;
+  }
+  & > div:nth-child(1) {
+    position: relative;
+    pointer-events: all;
+  }
+}
