+++ src/components/elements/Animator/AnimatorLayout/AnimatorLayout.module.scss
@@ -1,0 +1,13 @@
+.animator {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  max-width: 100%;
+  max-height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+.animatorOuterImageContainer {
+  overflow: hidden;
+  flex-grow: 1;
+}
